Added HUD/UI Effects
Screen flashes when damage is taken and depending on amount of damage, effect will linger on the screen.
The used images are placeholders and can be replaced with proper assets.
Added a health controller that can be subbed out if need be.
Sound can also be played upon taking damage.
Damage can be replaced by boosts/powerups/pickups.
Health regenerates and makes effect fade.
Press F to test while in my scene.