Added an intro cutscene to the PVE level.
Currently, the intro cutscene is a prefab that located in ND> AS and is called Cutscene_LvL_1. Within the cutscene manager (found in the prefab) there is an object called 'ThePlayer', which in the copy of the level, is set to the 'LevelStreaming' object. This can be swapped with the player or anything else at any point. Please unpack the prefab and remove the Plane and Cube if necessary. This was created with the help from Kieran Moore, a student who engaged with this module previously. He sends his greetings to Kostas.