Polished off the old PvE level ready for playtesting
Put the terrain back into the level, as it had previously been moved out of the way so we could work on the level, this way players won't fall through the floor of the map.
Added blank walls as placeholders around the map so players can't fall off the edge of the map.