Finished level streaming, and edited terrain
Finished setting up level streaming for UX and GD. All departments now how scenes in the level.
Also edited the walls around the map to encase the map further, and remove dead space around the map.
Finally, finished editing the terrain so that it covers part one and part two of the PvE level