Tutorial level polishing
Fixed an OOB when travelling into a specific teddy at high speeds. Added a few setdressing elements to the level, including distant walls to hide the default skybox. Rebaked lighting, for new objects since the first bake. Rotated car spawn point so that at least the car faces the right way at the start, if not the camera.